Try to ensure that changes to mmapped files are written back before close returns.
This is a speculative fix for NFS. The NFS FAQ says (see http://nfs.sourceforge.net/ - D8)
that closing a file is not enough to guarantee even close-to-open cache coherency and one
needs to call msync(MS_SYNC) to achieve that. Currently the only way to do that under Java
is to call FileChannel.force.